refactor: split global search result sources
